Fix to pppp drop-off and file-sending failures
In-progress fork of ankerctl to fix some PPPP connection leaks, video streaming cutouts of the printer and connection, file handling for sending the prints to the printer (resilient to unstable network and rate limited to prevent oversending).

Video streaming was removed as it was not implemented correctly and needs a separate stream handler to function properly without affecting other functionality.